Update ai framework
Changelog
- Move constants to _settings.js
- Fix bootstrap controller
- Add global spawnPriorityQueues
- Add spawn specific spawnQueues
- Add spawn specific spawnPriorityQueues
- Whitespace cleanup
- Added debug message

Spawning algorithm:

A spawner will always attempt to spawn a creep in this order

- Spawn specific spawnPriorityQueue*
- Spawn specific spawnQueue
- Global spawnPriorityQueue*
- Global spawnQueue

(* abort all steps if queue > 0, independent of possibility to spawn)

Please note that it is possible that -because of the global spawnPriorityQueue
can be cleared after an attempt is made by the spawner-, the spawner will
spawn nothing even when it should clear the global spawnQueue.

This issue might or might not get fixed later.